Adjust and refactor tests
