+++ test/data.json
@@ -1,0 +1,9 @@
+{
+  "sass": {
+    "withoutComments":
+      "$gray-base: #000 !default;\n$gray-darker: lighten($gray-base, 13.5%) !default; // #222\n$gray-dark: lighten($gray-base, 20%) !default;  // #333\n$gray:  lighten($gray-base, 33.5%) !default; // #555\n$gray-light:  lighten($gray-base, 46.7%) !default; // #777\n$gray-lighter:  lighten($gray-base, 93.5%) !default; // #eee",
+    "withComments":
+      "$one: 123;\n$x: $one;\n// $y: $two; // ERROR - $two not existed, but it's commented",
+    "indented": "$one: 123\n$x: $one\n"
+  }
+}
